HashSet pour les entrées clavier et un peu de refactoring

+++ src/main/java/engine/GameEngineGraphical.java
@@ -65,7 +65,7 @@ public class GameEngineGraphical {
 		int frames = 0;
 
 		long lastTimer = System.currentTimeMillis();
-		double dt = 0;
+		float dt = 0;
 
 		// boucle de game
 		while (!this.game.isFinished()) {
