+++ src/main/java/engine/IGame.java
@@ -10,8 +10,9 @@ public interface IGame {
 
 	/**
 	 * methode qui contient l'evolution du jeu
+	 * @param dt
 	 */
-	public void evolve(double dt);
+	public void evolve(float dt);
 
 	/**
 	 * @return true si et seulement si le jeu est fini
